requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
